Update patch set 1
Patch Set 1:

Sorry!

When I was trying to enable intrinsics with CanCall for ART_USE_READ_BARRIER=true, I encountered this issue, which introduces several run-test failures.

Obviously, this offset is mis-interpreted here, and is not consisted with other cases. But this is not a problem for concurrent source tree, and no appropriate regression test.

Patch-set: 1
Label: Verified=0