Update patch set 1
Patch Set 1:

> Aart, the problem is that you can hoist instruction from the loop
 > to pre-header only if you proved that this instruction will be
 > executed at least once because pre-header will be executed for sure
 > while instruction in the loop might be not.

I understand that, but if the array is proven within bounds (and not null), preloading the reference is not semantically incorrect as the commit message implies, although it may be less efficient for non taken loops. For unit stride references in an outer loop, this typically implies the loop is known to be taken, or the reference is moved within an explicit taken test. For constant subscripts, we may load "too much" for non taken loops, but never unsafe.

But yes, please submit your test so I can see what you worry about (note that the test should have been part of the original CL of course).

Patch-set: 1