Update patch set 1
Patch Set 1:

> Aart, just imaging that array access in a loop is guarded by if to
 > access an array only if there is no out of bound condition. BCE
 > will be smart enough to prove that array access is safe. But if you
 > move array access to pre-header without if that may lead to the
 > crash. So my first statement that to hoist invariant should be
 > executed at once is incorrect actually, we should prove that it
 > will be executed on the first iteration.

How can it crash? If the bounds check is eliminated, we know that the array element exists, even if the inner loop is not taken. And really, I understand what you are saying about having to go through the loop at least once. It is just that dynamic bce enforces that information, while static bce knows that the array element exists.

I think what you guys are overlooking is that this transformation *only* kicks in when inner bounds check is eliminated prior to hosting the array reference, to make it invariant and, thus, hoistable.

So for a[i][j] in nested loop, hosting a[i] only works if we bce its bounds check (independent of what we do with a[i][j] bounds check)

for
   for  
      
       (11) boundscheck ..       <-- if this stays, hoist does not happen
       (12) arrayget ... (11)     a[i]
       ...
       (15) boundscheck ..
       (16) arrayget (15)     a[i]j[]
  
    enddo

both bounds check are eliminated by doing some deopt tests in a block that is protected by the inner trip count. Then, the array references are hoisted into the same block, also protected by trip count. Lastly, there is a synthetic phi node after the block to merge the array read that exists with one that does not.

Please show me your example! Even if we overlooked something, this CL is not the way to fix it since it will regress a lot of performance.

Patch-set: 1