Update patch set 4
Patch Set 4:

ok, it seems that dominate all back edges means that instruction will be executed for all iterations possibly except the first one. For the movement of an array access I see the following side effects: NPE, IOBE and violation of lock semantic. The first two seems ok. I hope that the last one is only checked by BCE. Anton, Aart can you ensure in it? If it is ok Aart's solution seems ok to me. The worst thing happens then is redundant memory access for the loop which does not execute at least once and this is a low probability.

Patch-set: 4
Label: Verified=0