Update patch set 1
Patch Set 1:

What I explained to Sebastien is that those variables are not mandatory and this group of CL is ensuring that the jack script installed while building an android tree is aware of the Android tree defaults (version and arguments).

This CL is just to ensure that the environment variables are correctly transmitted to the script when they exists.

Patch-set: 1