Update patch set 5
Patch Set 5:

Sorry, I've no idea about the timeline. The bionic change is, imho, ready - but it has been sitting there for a month and must be updated to avoid merge conflicts. There are also several build system changes safestack depends on. And it also depends on b/27729263#12.

The current implementation has minor memory issues, see b/28909976. They are on the magnitude of 1Mb RAM device-wide, and will soon be fixed in upstream llvm.

IMHO we don't need a special bot for SafeStack. When it is enabled, it will be tested on all existing bots. We can revisit this question if we decide to keep it disabled for a significant amount of time.

Patch-set: 5
Label: Verified=0