Update patch set 4
Patch Set 4:

> Alexey, the test-art-target-gtest-relative_patcher_mips_test32 is
 > failing when executed on MIPS64r6 emulator.

Thanks for catching that!

It looks like we have a bug in MipsInstructionSetFeatures::FromVariant(). Neither "mips32r2" nor "default" clear the r6 flag and so the test that's supposed to exercise R2 paths in the compiler is taking R6 paths.

I'm going to dive in MipsInstructionSetFeatures::FromVariant() and see how I can fix it.

I can disable this test if it speeds up review/acceptance.

Patch-set: 4
Label: Verified=0