Update patch set 6
Patch Set 6:

> Very unhappy with the commit message. It mentions in no words to
 > strong change to force-stash into the dex-cache at find-class time
 > - which is the *center-point* of making this change work.

That very much looked like an oversight. We are often stashing things in the dex cache eagerly to speed up future lookups. Any reason you might think it could have been intentional not to do it?

 > 
 > (Plus it restricts us into the future. I *want* to inline across
 > dex files.)

No. This change is a bug fix. The restriction was, and is still there.

Patch-set: 6