Update patch set 1
Patch Set 1: Code-Review-1

This change won't work as-is.

If you increase the frame size that much, there's a chance we cannot convert a segfault because of stack guard into a stack overflow.

We may have to change the makefile to pass yet another define to increase the guard size (kStackOverflowImplicitCheckSize).

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1