Update patch set 6
Patch Set 6:

> I think there's tons of other cases where we'd like to load from
 > memory, and that requires some thoughts of having a unified way of
 > doing it. See my comment on PS5.

I talked to Nicolas offline, his -1 is about the current state of the CL. If you have the time and determination to implement this with the "emitted at use site" approach, we will be happy to accept it. It can be used as a blueprint for other situations where we can emit instructions that work directly with the heap.

Patch-set: 6
Label: Verified=0