Update patch set 4
Patch Set 4:

(1 comment)

Is there a way to use wait/notify on the object lock instead?

We basically seem to be reimplementing a poor man's condition variable here in order to save space by using thinlocks in the fast case.  But our original solution had a lost wakeup issue, and this isn't particularly clean.  It may still be the best we can do.

Patch-set: 4
Label: Verified=0