Update patch set 3
Patch Set 3:

> Nicolas and I were unhappy when Intel tried to introduce this
 > before, FYI.

Care to elaborate on the objections a bit? 

I suspect Nicolas finds the required control code too intrusive but, referring back to the design doc, having support for level 1 (methods) and level 2 (optimization phases) is really not much. Level 3 (steps within the optimizations) may become a point for discussion, but it is my humble opinion that even that code is justified by the easy of debugging that follows.

Perhaps another argument is that we have some strange forward dependencies in our optimizations, so apparently we cannot skip parts of one phase since later phases rely on that. We are still exploring what these are exactly, but without knowing all details, this seems to smell a bit anyways...

Patch-set: 3