Update patch set 3
Patch Set 3:

> The argument is against you introducing *another* set of parameters
 > controlling the compilation of methods, and that's what you're
 > doing. You introduce *two* new parameters to change compiler
 > behavior, in an IMHO super ugly way ("up to method-number" is
 > something really horrible, because it isn't really well-defined).
 > That *could* be rolled into the existing parameter, if that is
 > generalized.

Citation needed for "super ugly".

The # method approach is reasonably well-defined IMHO (by argument it worked so well in my past experience). I will have to study the switches you refer to a bit better, but those seem to have the disadvantage that you need to know the method names in advance, or somehow provide them, which make bisection harder.

But if the only argument is to reduce switches, let us study what is there and come back to you.

Patch-set: 3
Label: Verified=0