Update patch set 1
Patch Set 1: Code-Review+1

I'm not sure that it is a best way from design point of view. Might be it is better to make a subclass of ThreadPool which support only one Task in a queue of some kind and use this thread pool in JIT. So far I come up with this solution.

Patch-set: 1