Update patch set 1
Patch Set 1:

(5 comments)

My only thought is that in the past Nicolas wanted to have new bytecodes guarded behind a command line flag (-Xexperimental). If that's still the case it may not be wise to remove that support.

If that's no longer the case, you may still have an uphill struggle trying to simultaneously add correct support across the full breadth of art and fix all the regressing art tests due to adding new bytecodes.

Patch-set: 1
Label: Presubmit-Verified=0