Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > This is still super WIP.
 > 
 > My plan is to try to make this as minimal as possible by not really
 > implementing many of the jvmti functions and just getting the
 > loading stuff working.

I think the JVMTI parts should be separated out. Merge the minimal ART agent part first (with a gtest), then put the JVMTI connector on top of that.

The CL is not doing a lot but a s***-load of lines already. I'd feel a lot more comfortable reviewing smaller chunks.

Patch-set: 4