Update patch set 1
Patch Set 1:

> > I think we could do a memory "test" for 1 less instruction, maybe
 > > more code size though.
 > 
 > That would reorder the loads.

Well, we could do that without load reordering, given that the reference load does not modify flags. And we could use the "TEST r/m8, imm8" instead of reading the whole lock word to keep the code size low. The unpoisoning gets in the way though as it can change flags, but we could solve that by moving it after the branch and duplicating it in the slow path.

Patch-set: 1