Update patch set 1
Patch Set 1:

This is an initial version of the script. It worked fine during my manual testing, but it doesn't handle pass dependencies. I thought we could get away with simply pinning a number of mandatory passes but it doesn't seem to be enough.

It also retrieves debug output from device in a hacky way, via logcat. This is something we will be able to avoid once I implement a more legitimate mechanism for redirecting error output from logcat to stderr in dalvikvm on device.

Please take a look. :)

Patch-set: 1
Label: Verified=0