Update patch set 1
Patch Set 1:

I talked to Richard offline about this, essentially what we want is the reverse of logwrapper.

logwrapper is stdout/stdin -> logcat redirect.

We want logcat -> stdout/stdin redirect.

---------

So what if we have some kind of logredirect general tool that worked for any android native process ?

  adb shell logredirect dalvikvm ...

?

Or it could even be an environment variable

  adb shell ANDROID_LOG_REDIRECT=... dalvikvm

------------

So the other trouble is that because dalvikvm spawns sub-processes we'd want any children of that PID to also redirect logcat. This is pretty much the sole reason for why we can't just the regular logcat command and grep.

Patch-set: 1
Reviewer: Igor Murashkin <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0