Update patch set 35
Patch Set 35:

> The most important question was not addressed: Do we still support
 > devices with non-atomic LDRD?

It hasn't been answered because it is not the most important right now.

All supported Nexus devices are wired up with the config for atomic ldrd. Whether someone out there is still using the codebase with an older ARM device, who knows. So we have time to clean this up.

I discussed this with Christina and Mathieu before, and the solution we decided on was to provide two implementations (the other one the direct slow-path), let the runtime understand instruction-set-features (something it doesn't do right now), and then initialize the entrypoint accordingly.

I don't think there's any pressure as long as Christina or Mathieu don't forget about it.

Wrt/ to dropping support overall, I don't think that's up to you or me...

Patch-set: 35