Update patch set 2
Patch Set 2:

> Does the vdex file need to be treated separately from an oat file?
 > Could you encapsulate the notion of vdex inside OatFile entirely,
 > so that, even if vdex is a separate file on disk, logically it is
 > written and loaded as part of writing and loading the oat file?
 > That way, perhaps, you wouldn't have to expose vdex and
 > vdex_filename in so many different places in the runtime, such as
 > the OatFileAssistant.

OAT being an extension of VDEX is exactly how I'm thinking about it (and an image file extending OAT). It makes VDEX completely transparent to other components, but on the other hand we end up converting the OAT location into a VDEX location everywhere. I guess we don't care about being able to specify an arbitrary VDEX+OAT pair, do we? I was hoping it could be an entry in the OAT header but that's tricky...

Patch-set: 2