Update patch set 1
Patch Set 1:

I don't really mind this but last time I checked base was also shipping on windows and mac (and is it meant to work with vc/gcc or only clang)?

So is there instructions how to test the build and the correctness for all the other platforms?

And what are the criteria for us submitting changes to macros.h?

Patch-set: 1
Label: Verified=0