Update patch set 2
Patch Set 2:

Is the issue with the ART logging the stack backtraces or something else?  Aart and I would like to disentangle dexdump and friends from libart also and it would be nice to have a single solution to the logging dependency.

Patch-set: 2
Reviewer: David Sehr <1103743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0