Update patch set 1
Patch Set 1: Code-Review-1

What is the code expansion for this? What's the load-time impact.

Why not punish **** code by just not marking the class as verified in the oat file? That doesn't need a rev of the oat file, and has no size or load time impact.

Patch-set: 1
Label: Code-Review=-1