Update patch set 4
Patch Set 4:

So this is for optimizing post-loop access of instructions defined in a do/while loop?

Nice optimization. However, the way it's implemented looks very unnatural to the algorithm. I think ideally I would remove the spilling from the resolution, but just have the linear scan understand when to spill (for example it's the first interval with a stack slot). This will generalize better the optimization of where to best spill.

Patch-set: 4
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>