Update patch set 1
Patch Set 1:

> But that's a bad from a separation-of-concerns standpoint. If the
 > disassembler needs to be reverted, the test fixes shouldn't be
 > reverted.

I think in this case there are two types of test fixes - one concerns the test issues that we have identified, and the other is about the slight differences in output (including the ARMv7 vs. ARMv8 differences I have been discussing with Vladimir). If you revert the disassembler patch without reverting the second type of test fixes, then the tests will start failing. That's why I am perfectly fine with making the first type of fixes in a separate patch (in fact I prefer it), but keeping the second type in the current one. Of course, if you'd rather have all test fixes in a second patch, I can do that too.

Patch-set: 1