Update patch set 4
Patch Set 4:

> > I noticed that the JNI compiler was emitting VPOP/VPUSH for the
 > > transition (and the old disassembler just thought it was ???) ,
 > > maybe outside of the scope of this CL, would it be possible to
 > add
 > > a test for it?
 > 
 > Yes, but I'd rather do it separately. Do you have a more concrete
 > idea about what the test would look like? Perhaps it should be an
 > addition to 508-checker-disassembly.

Technically there should be a gtest similar to the assembler tests that tests for each instruction that may be emitted that it can be disassembled. Now that sounds like a lot of work... :-)

For the time being, running oatdump over core images and making sure there are no unknown instructions would be a huge step already, I think.

Patch-set: 4