Update patch set 6
Patch Set 6:

I have removed the ifdef.  Running "time dex2oat" for both current and the new version did not reveal any sensible regression - the wall time for that is in the range of 80 to 100 msec for both.

Are we good to go?

Patch-set: 6
Reviewer: Alexey Alexandrov <1122300@85c56323-6fa9-3386-8a01-6480fb634889>