Update patch set 1
Patch Set 1:

In case it's not immediately obvious, "u" means unordered, and the changed comparisons were comparing for the condition that was a union of whatever the condition needed to be checked and unordered. This caused NANs (unordered) to be mishandled. The code should've been the same on R2 and R6 in this respect, we just had a typo and missed some MIPS32R6 testing.

Patch-set: 1
Label: Verified=0