Update patch set 4
Patch Set 4: Code-Review-1

> > Why? There's no actual commit message to motivate this.
 > 
 > See "OTA friendly vdex files". Do you want me to do add a short
 > summary of that doc in the commit message?

Each commit message should explain what's going on (especially considering that the document might be outdated/updated/inacccessible).

I do not remember why putting quickened bytecode into vdex would make it OTA friendly. In fact, on the face of it, it is *not* OTA friendly, as quickened bytecode relies on internal layout of the boot image and as such could be invalidated by an OTA. So an explanation seems required.

Patch-set: 4
Label: Code-Review=-1