Update patch set 2
Patch Set 2:

(1 comment)

Thanks for the feedback!

> Please add an actual commit message motivating your change

Done.

> What's the reason the change is needed?

AIUI, if you call open() with O_CREAT, you need to pass in a mode parameter. Is this not correct on Android?

The reason I care is that calling open with O_CREAT + no `mode` argument used to be a compile-time error. When we swapped to compiling the world with clang, that stopped being the case. We're trying to make it a compile-time error once again, but if it's actually sane to call open with O_CREAT and without a third param on Android, I'm no longer fully convinced that it should be an error. :)

Patch-set: 2
Reviewer: George Burgess IV <1096946@85c56323-6fa9-3386-8a01-6480fb634889>