Update patch set 2
Patch Set 2:

I could push the IsInBlock() check into the iterator used by BCE. This works, since the next field is not changed on replacement. Or I could do a callback when each instruction is replaced and adjust the current (no need to specifically do that on the next in that case). Before coding this out, please let me know your preference.

Patch-set: 2