Update patch set 4
Patch Set 4:

> (5 comments)
 > 
 > I think I don't understand the meaning of --update-vdex well enough
 > to offer an opinion on whether the logic in this CL is correct. Can
 > you clarify what --update-vdex means?

Sure, and just to confirm, I haven't changed the meaning since the discussion you, Calin and I had. So here goes.

With the introduction of vdex, we now have two ways to handle a "dex2oat_needed": 1) it's needed because the oat file is out of date wrt the boot image, and 2) it's needed because there is a better compiler filter.

In the situation of 1), we can change directly the existing vdex file and avoid creating a new one. In the situation of 2) we cannot change the existing vdex file, as running apps may be using it.

So --update-vdex is the name we chose during our discussion to notify dex2oat it is in situation 1), where it can update directly the existing vdex file.

Hope this clarifies?

Patch-set: 4