Update patch set 3
Patch Set 3:

> > > Please break this up so it's either in the case that reverts
 > will
 > > > be necessary (like last Fri).
 > >
 > > Do you mean break things up between archs?
 > 
 > Yes.

We haven't had this rule really, and sometimes you have to have all archs implemented for a feature in one change (for example adding a new HInstruction, and doing instruction set switches isn't ideal).

I personally like the approach of doing one as a foundation step and not overwhelm the reviewer, and then follow-up with a change of all archs. But if the change is of reasonable size, having all in one change works for me too.

And about reverts: you could still revert this one change, even if it fails only for one arch. If you were to split things up, and there was a more general flaw in the optimization, you would need to revert all, which isn't great.

Patch-set: 3