Update patch set 2
Patch Set 2: Code-Review-1

There is a problem with this CL as-is.  When invoking `make` with a test target T responsible for running several tests (e.g. `test-art-host-gtest-codegen_test`, which runs the 32- and 64-bit versions), if at least one test is failing (meaning the code in `ART_TEST_FAILED` is executed), then the summary of skipped/passed/failing tests won't be printed at the end.  This is because the failing test is a prerequisite (dependency) of the rule building T, which displays the summary as an action of the rule, using `ART_TEST_PREREQ_FINISHED`. In other words, `ART_TEST_PREREQ_FINISHED` precisely requires *all* the prerequisites of the rule where it is used as an action *not* to fail.

I don't think it is desirable to lose this feature (printing the test summary whatever the result of the tests).

Maybe we could try and make `ART_TEST_FAILED` smarter to detect whether it is run in a context where `ART_TEST_PREREQ_FINISHED` is expected to be run eventually.

Patch-set: 2
Reviewer: Roland Levillain <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0
Label: Code-Review=-1