Update patch set 3
Patch Set 3:

Would it make sense to have some attributes that are computed for methods and retained in some way outside the inliner but not in the HLoopInformation?  The no-exit-loop condition is invariant of call site and it's a pity to compute it over and over.  And I'm not sure many optimizations downstream are going to use the information once it is computed.

Patch-set: 3
Label: Verified=0