Update patch set 1
Patch Set 1:

(1 comment)

I guess this case happens when the class loader is unreachable but we are allocating an instance of a class that belongs to that class loader concurrently with a sticky GC that would unload/deallocate the class.

It seems we don't keep the class alive on the allocation path or the dex cache type array that we'd load the class from won't keep the class alive.

Why isn't this a problem for the partial/full case?

Why isn't this a problem for other non-allocation operations like instanceof or a static call?

Anyhow, not unloading classes for sticky sounds fine to me.

Patch-set: 1