Update patch set 2
Patch Set 2:

Mingyao made a good observation that something in LICM is not quite right (independent of this change). Something like the code below prints 0, not 1 because we ignore the visible side effects of the field store (only looking at can throw properties).

I will file a bug on that independent of this change.


public class M {

  static int a;

  static void foo(int[] arr) {
    int j = 0;
    do {
      a = 1;
    } while (j < arr.length);
  }

  public static void main(String[] args) {
    a = 0;
    try {
      foo(null);
    } catch (Exception e) {
    }
    System.out.println("a=" + a);
  }
}

Patch-set: 2