Update patch set 2
Patch Set 2:

> > What's the context?
 > >
 > > And is that necessarily true? There's no class in openjdk or in
 > the
 > > framework that would fail to initialize due to user/app input?
 > 
 > (verified to workaround gerrit bug)
 > 
 > OK, got the context, from a build breakage thread, that you and
 > Calin are investigating, thanks!
 > 
 > But do you want to keep this change? Can we ensure what I've
 > mentioned in my previous message?

I think we want to keep this, but maybe weaken to a DCHECK (which I had before).

App classes should not be able to interfere with boot classpath initialization. If that's the case, I think that is a bug in how the code is written.

Patch-set: 2
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>