Update patch set 4
Patch Set 4: -Code-Review

Orion - I was having a chat about this with our kernel folks. Do you have any stack trace that points at this issue? 

It would be helpful to see what the kernel thinks is wrong - by looking at the ESR. If you or the Samsung folks have one it would be helpful in helping identifying what is wrong. Also, what is the kernel version that you see this issue on? Can I reproduce it in any way?

As for the ARM ARM reading it carefully this should be permitted, but it might be that the Linux kernel has additional constrains.

As for big.LITTLE systems, they are probably more prone to hit this issue due to their asymmetry, but I do not think this is a big.LITTLE specific issue (unless it is a CPU errata).

Patch-set: 4
Reviewer: Serban Constantinescu <1072549@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0