Update patch set 1
Patch Set 1:

(1 comment)

I'd like to check if this seems like a good thing to check and should it be more restrictive (not just checking constants).
The test that fails when checking everything 592-checker-regression-bool-input implies that similar check was present in GraphChecker and is now removed because it was not trivial to do it there.
Is this a better place to do it?

Patch-set: 1
Reviewer: Branislav Rankov <1125523@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0