Update patch set 2
Patch Set 2:

We were unable to auto-submit your CL due to the following error: Failed to submit 3 changes due to the following problems:

Change 306785: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.

Change 307663: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.

Patch-set: 2