Update patch set 4
Patch Set 4: Code-Review+1

TLDR: I think we should merge this, as it is an improvement. I think there is another location at risk in MoveLocation; I opened a ticket for it.

Details:

I took some time to look at the `ParallelMoveResolverARM64` and `CodeGeneratorARM64::MoveLocation`.

After Roland's patch I think the case where we swap two stack slots is safe because:
- Today there are no calls to `ParallelMoveResolverARM64::EmitMove` inside a
  `UseScratchRegisterScope`. This is unlikely to change much. Parallel moves are
  not widely used and the code using them does not change often.
- From what I understand, the `ParallelMoveResolverNoSwap` will require at
  most one scratch register at a time to break cycles. (See comment in
  `ParallelMoveResolverARM64::PrepareForEmitNativeCode`)
- The code generated in `CodeGeneratorARM64::MoveLocation` should not require
  more than one scratch register on the `MacroAssembler` side. If that changes
  an assert will fire.

However I think there a crash could occur if
- the destination is a stack slot and a source is a constant (different from 0)
- and we are emitting code for a move while the parallel move resolve already acquired a scratch core register
- and the stack offset at which we are storing cannot be encoded in the STR instruction.

STR has a good range on AArch64 (16k/32k for 32bit/64bit), but STUR does not (-256 to 255). So if 4 bytes and 8 bytes values are stored to respectively 4 bytes and 8 bytes aligned on the stack it should be extermely rare. But if (and I think that is the case) slots for 8 bytes values are just two contiguous 4 bytes slots without alignment guarantee, then we are at risk. This should still be quite rare though. I have not seen many cases when we needed a move to store a constant on the stack.

I will open a ticket internally to look at this.

Patch-set: 4
Label: Code-Review=+1