Update patch set 5
Patch Set 5:

> The CL https://android-review.googlesource.com/#/c/268914/ will be
 > abandonded in favor of this CL.
 > 
 > Calin, any plans to merge this one?

Hi Serguei. Yes, it is still in plans. The reason for the delay is that this depends on https://android-review.googlesource.com/#/c/307923/ which depends on https://android-review.googlesource.com/#/c/306923/. The last one is rather tricky but I think I'll be able to submit it soon and unblock he chain.

Patch-set: 5