Update patch set 1
Patch Set 1:

A separate, but related to a redesign, issue is that it is easy to introduce bugs in the spirit of malloc/free bugs with the current API. It is hard tool around this as register an allocation and freeing an allocation API don't produce and consume a cookie, a pointer in the case of malloc/free. As the behavior of over freeing is to throw an exception, this can lead to production crashes.

Patch-set: 1