Update patch set 1
Patch Set 1:

> > > Yes, we will look into this right away.
 > >
 > > Thanks!
 > 
 > Have found the exact reason, will upload the fix very soon.

Mathieu should have commented yesterday that we discussed this, and think that the cleanest solution is to refactor the MacroAssembler CallingConvention connection - use the assembler instead of a Create function. That way the assembler can be a scope, and use the VIXL support for using scratch assemblers.

Patch-set: 1
Label: Verified=0