Update patch set 1
Patch Set 1:

> Please add more rationale. Disabling the poisoning there prevents
 > finding potential moving GC bugs. Maybe there's a jvmti reason you
 > want this, but it would be great to avoid it.

It doesn't, that's the point. If moving GC is disabled, there are no bugs with raw pointers, because objects do not move.

There's no way to avoid this. If an input is GC-safe (a jobject), but I need acccess in a suspended scope, right now there's no way to access this:

* decode before -> SuspendAll is poisoning point
* decode after -> impossible, decoding needs mutator lock

The following is perfectly safe:

 DisableMovingGC();
 MakeRawPointer();
 SuspendAll():
 UsePointer();
 ResumeAll();
 EnableMovingGC();

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>