Update patch set 1
Patch Set 1:

OK, I didn't realize IsMovingGCDisabled was from an explicit action, and not the GC being able to move or not.

But the fact that DisableMovingGC/EnableMovingGC are operations done regularly still means we may miss moving GC bugs that are also called in a different context.

So how about adding a Enable/DisablePoisoningObjPtr?

Patch-set: 1
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>