Update patch set 1
Patch Set 1:

Thanks for the feedback. 

I looked into this more. And as it turned out, this idea doesn't seem to have an upside because there are rarely situations in practice where we just want to do a clinit check without needing the updated reference (via the read barrier) and there isn't much point in loading the class without a read barrier and checking its status.

Among the instructions that need a clinit check, StaticFieldGet/NewInstance need an up-to-date class reference anyway (which needs a read barrier) and InvokeStatic uses the resolution stub to eliminate the clinit check after the very first invocation.

Patch-set: 1
Reviewer: Hiroshi Yamauchi <1022530@85c56323-6fa9-3386-8a01-6480fb634889>