Update patch set 5
Patch Set 5:

> (8 comments)
 > 
 > (Sorry) I also forgot you need to update:
 > - Instrumentation::deoptimized_methods_.
 > - debugger.cc: gBreakpoints.
 > 
 > Also, +Seb for the liveness of JDWP::MethodId (that just contain
 > the ArtMethod pointer), do we need to worry about those?

That's a good question. AFAICT the MethodID is a unique reference to the method so that the debugger client can reference it. It must not change during the debugging session. In ART, this is just the address of the corresponding ArtMethod*. 

Now, taking obsolete methods into account, I can't say if the MethodID must remain the same or not. Though we still need to the old MethodID to implement the Method.IsObsolete command. I don't know what is the behavior of the debugger client when methods are updated. Maybe it look for the MethodID of the new method by issuing a ReferenceType.Methods command to the reference type. You could run experiments on the RI to understand what happens.

There is more information from the command VirtualMachine.RedefineClasses which allows to update a class:

"
Installs new class definitions. If there are active stack frames in methods of the redefined classes in the target VM then those active frames continue to run the bytecodes of the original method. These methods are considered obsolete - see IsObsolete. The methods in the redefined classes will be used for new invokes in the target VM. The original method ID refers to the redefined method. All breakpoints in the redefined classes are cleared.If resetting of stack frames is desired, the PopFrames command can be used to pop frames with obsolete methods.
"
(from https://docs.oracle.com/javase/8/docs/platform/jpda/jdwp/jdwp-protocol.html#JDWP_VirtualMachine_RedefineClasses)

For the breakpoint part, I think that the redefined method should not inherit breakpoints set in the original method: "All breakpoints in the redefined classes are cleared". 


(Note that we do not support the StackFrame.PopFrames command.)

Patch-set: 5
Label: Verified=0